refactor: consolidate query output files (18 → 4)
Merge 18 query_*.py CLI files into 4 focused modules:

- query.py — orchestration: frontdoor routing, plan execution, helpers
- query_output.py — all output formatting, delivery, list, stream, summary
- query_stats.py — grouped stats, SQL stats, profile stats, structured
- query_semantic.py — semantic slice and semantic stats

Kept query_actions.py as-is (already focused).

Net: 18 files → 4 files, -144 lines.
