+++ polylogue/cli/click_app.py
@@ -20,7 +20,7 @@ from polylogue.cli.click_option_groups import apply_query_mode_options
 from polylogue.cli.formatting import announce_plain_mode, plain_forced_by_env, should_use_plain
 from polylogue.cli.machine_main import extract_option as _extract_option
 from polylogue.cli.machine_main import run_machine_entry
-from polylogue.cli.query_frontdoor import QueryFirstGroupBase, handle_query_mode
+from polylogue.cli.query import QueryFirstGroupBase, handle_query_mode
 from polylogue.cli.types import AppEnv
 from polylogue.logging import configure_logging
 from polylogue.ui import create_ui
