--- polylogue/cli/query_output_formatting.py
@@ -1,95 +1,0 @@
-"""Formatting helpers for CLI query output."""
-
-from __future__ import annotations
-
-import json
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from polylogue.cli.types import AppEnv
-    from polylogue.lib.models import Conversation
-
-
-def format_list(
-    results: list[Conversation],
-    output_format: str,
-    fields: str | None,
-) -> str:
-    """Format a list of conversations for output."""
-    from polylogue.rendering.formatting import _conv_to_dict
-
-    if output_format == "json":
-        return json.dumps([_conv_to_dict(c, fields) for c in results], indent=2)
-    if output_format == "yaml":
-        import yaml  # type: ignore[import-untyped]
-
-        return str(yaml.dump([_conv_to_dict(c, fields) for c in results], default_flow_style=False, allow_unicode=True))
-    if output_format == "csv":
-        from polylogue.cli.query_summary_output import conversations_to_csv
-
-        return conversations_to_csv(results)
-
-    lines = []
-    for conv in results:
-        date = conv.display_date.strftime("%Y-%m-%d") if conv.display_date else "unknown"
-        raw_title = conv.display_title or conv.id[:20]
-        title = (raw_title[:47] + "...") if len(raw_title) > 50 else raw_title
-        msg_count = len(conv.messages)
-        lines.append(f"{conv.id[:24]:24s}  {date:10s}  {conv.provider:12s}  {title} ({msg_count} msgs)")
-    return "\n".join(lines)
-
-
-def render_conversation_rich(env: AppEnv, conv: Conversation) -> None:
-    """Render a conversation with Rich role colors and thinking block styling."""
-    from rich import box
-    from rich.markdown import Markdown
-    from rich.panel import Panel
-    from rich.text import Text
-
-    from polylogue.ui.theme import THINKING_STYLE, provider_color, role_color
-
-    console = env.ui.console
-    title = conv.display_title or conv.id
-    pc = provider_color(conv.provider)
-    header = Text()
-    header.append(title, style="bold")
-    if conv.display_date:
-        header.append(f"  {conv.display_date.strftime('%Y-%m-%d %H:%M')}", style="dim")
-    header.append(f"  [{pc.hex}]{conv.provider}[/{pc.hex}]")
-    console.print(header)
-    console.print()
-
-    for msg in conv.messages:
-        if not msg.text:
-            continue
-        role = msg.role or "unknown"
-        rc = role_color(role)
-        is_thinking = msg.is_thinking
-        if is_thinking:
-            content = Text(msg.text[:500], style=THINKING_STYLE["rich_style"])
-            if len(msg.text) > 500:
-                content.append(f"\n... ({len(msg.text):,} chars)", style="dim")
-            panel = Panel(
-                content,
-                title=f"{THINKING_STYLE['icon']} Thinking",
-                title_align="left",
-                border_style=THINKING_STYLE["border_color"],
-                box=box.SIMPLE,
-                padding=(0, 1),
-            )
-            console.print(panel)
-            continue
-        try:
-            md = Markdown(msg.text)
-            panel = Panel(
-                md,
-                title=f"[{rc.label}]{role.capitalize()}[/{rc.label}]",
-                title_align="left",
-                border_style=rc.hex,
-                box=box.ROUNDED,
-                padding=(0, 1),
-            )
-            console.print(panel)
-        except Exception:
-            console.print(f"[{rc.label}]{role.capitalize()}:[/{rc.label}] {msg.text[:200]}")
-        console.print()
