+++ tests/unit/cli/test_query_exec_laws.py
@@ -19,7 +19,7 @@ from rich.console import Console
 
 from polylogue.cli.query import async_execute_query, project_query_results
 from polylogue.cli.query_actions import apply_modifiers, apply_transform, delete_conversations
-from polylogue.cli.query_helpers import no_results, summary_to_dict
+from polylogue.cli.query import no_results, summary_to_dict
 from polylogue.cli.query_output import (
     _output_summary_list,
     _render_conversation_rich,
@@ -29,7 +29,7 @@ from polylogue.cli.query_output import (
     output_stats_by_summaries,
     output_stats_sql,
 )
-from polylogue.cli.query_plan import (
+from polylogue.cli.query import (
     QueryAction,
     QueryExecutionPlan,
     QueryMutationSpec,
@@ -37,7 +37,7 @@ from polylogue.cli.query_plan import (
     QueryRoute,
     resolve_query_route,
 )
-from polylogue.cli.query_summary_output import output_stats_by_conversations
+from polylogue.cli.query_output import output_stats_by_conversations
 from polylogue.cli.types import AppEnv
 from polylogue.lib.models import Conversation, Message
 from polylogue.lib.query_spec import ConversationQuerySpec, QuerySpecError
